requirements.txt
appdirs==1.4.4
asgiref==3.3.4
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
autopep8==1.4.4
awsebcli==3.20.2
bcrypt==3.2.0
blessed==1.18.1
blinker==1.4
boto3==1.18.8
botocore==1.21.8
cached-property==1.5.2
cement==2.8.2
certifi==2019.3.9
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.3
click==7.1.1
colorama==0.4.3
configobj==5.0.6
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==3.4.7
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.14
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-invitations==1.9.3
django-paypal==1.1.1
django-registration==3.1.2
django-storages==1.11.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
example==0.1.0
filelock==3.0.12
Flask==0.12.5
Flask-Compress==1.4.0
Flask-SocketIO==2.9.6
future==0.16.0
gdbgui==0.13.2.0
gevent==1.4.0
greenlet==0.4.15
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
importlib-metadata==4.0.1
importlib-resources==5.1.4
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
netifaces==0.10.4
oauthlib==3.0.1
paramiko==2.7.2
parse==1.19.0
pathspec==0.5.9
pep8==1.7.1
Pillow==6.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
pygdbmi==0.9.0.3
Pygments==2.6.1
PyJWT==2.1.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==17.5.0
pyserial==3.4
python-dateutil==2.8.2
python-debian==0.1.32
python-decouple==3.1
python-engineio==3.12.1
python-socketio==4.5.1
python3-openid==3.1.0
pytz==2018.5
PyVimeo==1.1.0
pyxdg==0.25
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
s3transfer==0.5.0
SecretStorage==2.3.1
semantic-version==2.8.5
service-identity==16.0.0
simplejson==3.17.2
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.2.4
ssh-import-id==5.7
stripe==2.27.0
termcolor==1.1.0
texttable==1.6.4
tinydb==4.5.0
tuspy==0.2.5
Twisted==17.9.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uWSGI==2.0.19.1
virtualenv==20.4.7
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==0.16.1
zipp==3.4.1
mysqlclient
